html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-content-before-legend.html

<!doctype html>
<title>fieldset content before legend</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
fieldset {
  /* Paddings might have fractional values by default, and they can cause
     rounding differences between the test element and the reference element. */
  padding: 0;
}
</style>
<fieldset id=test>
  X
  <legend>legend</legend>
  Y
</fieldset>
<fieldset id=ref>
  <legend>legend</legend>
  X Y
</fieldset>
<script>
  test(() => {
    const testElm = document.getElementById('test');
    const refElm = document.getElementById('ref');
    assert_equals(testElm.clientHeight, refElm.clientHeight);
  });
</script>
