Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-content-before-legend-expected.txt

X legend
Y
legend
X Y

FAIL fieldset content before legend assert_equals: expected 34 but got 52
